+++ src/main/java/com/jcgfdev/deliveryapp/security/services/IConfirmationPathService.java
@@ -3,5 +3,5 @@ package com.jcgfdev.deliveryapp.security.services;
 import com.jcgfdev.deliveryapp.security.entities.ConfirmationPath;
 
 public interface IConfirmationPathService {
-    ConfirmationPath findById(Long id);
+    ConfirmationPath findById(Integer id);
 }
